modified for testing
