+++ package.json
@@ -4,7 +4,11 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "start":"node dist/index.js",
+    "dev": "nodemon --exec ts-node index.ts",
+    "build": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
+
   },
   "keywords": [],
   "author": "",
@@ -14,10 +18,13 @@
     "@types/node": "^20.10.4",
     "@types/sqlite3": "^3.1.11",
     "express": "^4.18.2",
-    "nodemon": "^3.0.2",
     "sqlite": "^5.1.1",
     "sqlite3": "^5.1.6",
     "ts-node": "^10.9.1",
     "typescript": "^5.3.3"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.0",
+    "ts-node-dev": "^2.0.0"
   }
 }
